refactor: assert empty or wrong URL and token, fixes the error
Assert empty or wrong URL and token, fixes the error
