+++ packages/cli/src/commands/styles/utils/figma.loader.ts
@@ -1,5 +1,7 @@
 import * as Figma from 'figma-api';
-import { FigmaFileParams, processFigmaUrl } from './figma.utils';
+import { FigmaFileParams } from './figma.utils';
+import { assert, regexSingleMatch } from './common.utils';
+import { logger } from '../../../logger';
 
 export const setupFigma = (key: string) => {
   const api = new Figma.Api({
@@ -13,9 +15,36 @@ export const setupFigma = (key: string) => {
 };
 
 export const loadFile = ({ url, token }: FigmaFileParams) => {
+
+  assert(url,'Figma URL Not Found, please run styles command again and provide the URL: ');
+  assert(token, 'Figma Token Not Found, please run styles command again and provide the token: ');
+
   const { getFileNode } = setupFigma(token);
   const { fileId, nodeId } = processFigmaUrl({ url, token });
+  assertFigmaData({ fileId, nodeId, token });
 
-  const input = getFileNode(fileId,[nodeId]);
+  const input = getFileNode(fileId,[nodeId]).catch(() => {
+    logger.error('Error loading Figma file');
+    process.exit(1);
+  });
   return input;
+};
+
+export type FigmaData = {
+  fileId: string,
+  nodeId: string,
+  token: string,
+};
+
+export function assertFigmaData(d: unknown): asserts d is FigmaData {
+  const { fileId, nodeId } = (d as FigmaData) ?? {};
+  assert(Boolean(fileId), 'Please enter the URL for a valid Figma file');
+  assert(Boolean(nodeId), 'Please enter the URL for a valid Figma file');
+}
+
+export const processFigmaUrl = ({ url, token }: FigmaFileParams): FigmaData => {
+  const fileId = regexSingleMatch(url, /\/file\/(.*)\//);
+  const encodedId = regexSingleMatch(url, /\?node-id=(.*)&?$/);
+  const nodeId = encodedId && decodeURIComponent(encodedId);
+  return { fileId, nodeId, token };
 };
